+++ NEWS.md
@@ -5,3 +5,4 @@
 # timefully 0.1.1
 
 * Included `complete_timeseries_datetime` function
+* Bug fix in `change_timeseries_tzone` function